+++ lib/dynamodb-wrapper.d.ts
@@ -9,6 +9,17 @@ declare interface IDynamoDBWrapperOptions {
 }
 
 declare interface IBatchWriteItemOptions {
+    [key: string]: IBatchWriteItemOption;
+
+    // DEPRECATED - moved to IBatchWriteItemOption
+    // still currently supported for backwards compatibility, but may be removed in the future
+    partitionStrategy?: 'EqualItemCount' | 'EvenlyDistributedGroupWCU';
+    targetItemCount?: number;
+    targetGroupWCU?: number;
+    groupDelayMs?: number;
+}
+
+declare interface IBatchWriteItemOption {
     partitionStrategy?: 'EqualItemCount' | 'EvenlyDistributedGroupWCU';
     targetItemCount?: number;
     targetGroupWCU?: number;