+++ lib/partition-strategy.ts
@@ -18,7 +18,7 @@ const BATCH_WRITE_MAX_ITEM_COUNT = 25;
  */
 
 export function getNextGroupByItemCount(writeRequests: DynamoDB.WriteRequests, startIndex: number,
-                                        options: IBatchWriteItemOptions): DynamoDB.WriteRequests {
+                                        options: IBatchWriteItemOption): DynamoDB.WriteRequests {
 
     if (startIndex < 0 || startIndex >= writeRequests.length) {
         return null;
@@ -72,7 +72,7 @@ export function getNextGroupByItemCount(writeRequests: DynamoDB.WriteRequests, s
  */
 
 export function getNextGroupByTotalWCU(writeRequests: DynamoDB.WriteRequests, startIndex: number,
-                                       options: IBatchWriteItemOptions): DynamoDB.WriteRequests {
+                                       options: IBatchWriteItemOption): DynamoDB.WriteRequests {
 
     if (startIndex < 0 || startIndex >= writeRequests.length) {
         return null;